requirements.txt
requests==2.26.0
python-dotenv==0.19.1
psycopg2-binary==2.9.1